requirements.txt
Django==1.10
django-custom-user==0.6
django-password-reset==0.9
psycopg2==2.6.2
